Several nontrivial changes
1. Moving *ModelRepn objects out of coek_model.hpp

These still needed to be in a header, but one that is only used internally.

2. Initial rework of the API for compact models.

Starting to all logic to allow variables to have expressions for lb, ub, value.

+++ coek/compact/coek_indexed.hpp
@@ -37,8 +37,9 @@ public:
         : data(other.data)
         {}
 
+    // The vector length is stored in the first element.  This limits the size, but this is OK for small vectors
     size_t size() const
-        { return *data; }
+        { return static_cast<size_t>(*data); }
 
     int& operator[](size_t i)
         { return data[i+1]; }
